refactor: move make_edges_df function to metadata so it can be used in multiple sections
