+++ src/readii_2_roqc/feature_extraction/index.py
@@ -7,60 +7,7 @@ from readii.io.loaders import loadImageDatasetConfig
 from readii.process.config import get_full_data_name
 from readii.utils import logger
 
-from readii_2_roqc.utils.metadata import get_masked_image_metadata
-
-
-def make_edges_df(mit_index: pd.DataFrame | Path,
-                   image_modality: str,
-                   mask_modality: str) -> pd.DataFrame:
-    """Create rows for matching images and masks from an imgtools autopipeline dataset index file. Each image may have multiple masks, so a row is created for each combination.
-    
-    Parameters
-    ----------
-    mit_index : pd.DataFrame | Path
-        DataFrame or Path to csv file containing the index data generated by running imgtools autopipeline. Can be from the index.csv or index-simple.csv. Must contain:
-
-        * 'filepath'
-        * 'PatientID'
-        * 'SampleNumber'
-        * 'Modality'
-        * 'ImageID'
-        * 'SeriesInstanceUID'
-        * 'ReferencedSeriesUID'
-    image_modality : str
-        Modality to filter the index by to get image rows (e.g. CT, MR)
-    mask_modality : str
-        Modality to filter the index by to get mask rows (e.g. RTSTRUCT, SEG)
-            
-    Returns
-    -------
-    edges_df : pd.DataFrame
-        DataFrame where each row contains metadata for an image and mask pair. 
-    """
-    if isinstance(mit_index, Path):
-        try:
-            # Check that the correct file name has been passed
-            assert mit_index.name.endswith("index.csv") or mit_index.name.endswith("index-simple.csv")
-            # Load the file into a pandas DataFrame
-            mit_index = pd.read_csv(mit_index)
-
-        except AssertionError:
-            message = "Expected imgtools autopipeline index file ending in 'index.csv' or 'index-simple.csv'."
-            logger.error(message)
-            raise
-    
-    # Get the image rows and mask rows from the MIT index and merge based on 
-    #   - mask's ReferencedSeriesUID == image's SeriesInstanceUID
-    #   - Matching SampleNumber
-    #   - Matching PatientID
-    edges_df = mit_index[mit_index.Modality == image_modality].merge(
-        mit_index[mit_index.Modality == mask_modality],
-        left_on=['SeriesInstanceUID', 'SampleNumber', 'PatientID'],
-        right_on=['ReferencedSeriesUID', 'SampleNumber', 'PatientID'],
-        suffixes=('_image', '_mask')
-    )
-        
-    return edges_df
+from readii_2_roqc.utils.metadata import get_masked_image_metadata, make_edges_df
 
 
 def get_base_index(dataset_config: dict,
